Cleanup: Pull inbox into components
